[tests] replace build_systems python tests as part of `make test-replace`
Summary:
The way to replace failing tests coming from build_integration_tests.py was not
very discoverable, and now that we have a target for precisely this action we
might as well use it.

The downside is that it's much slower to run `make test-replace` that way but
that's all the more incentive to migrate tests from python to Makefiles. Anyway
most of that time is spend in the buck test, which is going to be migrated
soon.

Reviewed By: jeremydubreil

Differential Revision: D4147121

fbshipit-source-id: 53689e9