[infer][java] mark all methods as defined
Summary: Mark native methods as defined so that the analysis generates a summary for those methods. When analyzing Java projects compiled with Buck, the summaries for the dependencies methods of are retrieved from the classpath. In this case, having access to the summary is useful to access the attributes of a callee when the callee is part of a, previously analyzed, Buck target.

Reviewed By: sblackshear

Differential Revision: D4141362

fbshipit-source-id: 75888c8