[python] add script report.py to post-process report.json
Summary:
This will help porting more Python code over to OCaml. Since the reporting step
uses a lot of Python libraries that would be a pain to rewrite in OCaml (eg,
syntax highlighting), keep this functionality in Python and make it possible to
call it from OCaml as a script.

Reviewed By: jberdine

Differential Revision: D4182832

fbshipit-source-id: fc83220