[threadsafety] analyze all the files and methods
Summary: Currently the thread safety checker neglects to analyze and files methods we don't want to report on. Like constructors and private methods, and classes where no superclass is marked ThreadSafe. For interprocedural analysis we want to analyze all these to get summaries, even if we don't report on them.

Reviewed By: jberdine

Differential Revision: D4226515

fbshipit-source-id: 7571573