[infer][java] remove the dead code manipulating the bootclasspath
Summary: The default `android.jar` used to compile Android code is just a list of empty method stubs that is used for the type checking. In order to improve the type information, we used to redirect the bootclasspath to another `android.jar` that was containing the method bodies and all the fields including the private ones. Since we no longer need the types in the models and the types in the libraries to match, this is now dead code.

Reviewed By: jberdine

Differential Revision: D4230510

fbshipit-source-id: 93417f3