[ondemand] simplify API by removing need to pass type environment
Summary: Noticed this when I was writing the documentation for the abstract interpretation framework and was curious about why `Ondemand.analyze_proc` needs the type environment. It turns out that the type environment is only used to transform/normalize Infer bi-abduction specs before storing them to disk, but this can be done elsewhere. Doing this normalization elsewhere simplifies the on-demand API, which is a win for all of its clients.

Reviewed By: cristianoc

Differential Revision: D4241279

fbshipit-source-id: 957b243