[thread-safety] don't report thread safety violations due to calls to constructors
Summary:
`o.<init>` cannot be called in parallel with other methods of `o` from outside, so it's less likely to have thread safety violations in `o.<init>`.
This diff suppresses reporting of thread safety violations for fields touched (transitively) by a constructor.
We can do better than this in the future (t14842325).

Reviewed By: peterogithub

Differential Revision: D4259719

fbshipit-source-id: 20db71f