Fixed problem with evaluation of sizeof which would give false positive
Summary:
When calling function g_realloc(gpointer mem,gsize n_bytes) one of the spec considers the case
whereby n_bytes is  zero. In that case g_realloc would return null.

If we call with sizeof(int), infer would compare sizeof(int) with zero. But the prover would fail to
understand that sizeof(int) != 0.
This diff fix this. We try to convert expression to constant when they can be converted (eg in case of sizeof).
The method currently make a partial set of conversion. This could be extended.

Reviewed By: jberdine

Differential Revision: D4166944

fbshipit-source-id: 3ec4fd7