[UI] add source context to reports in bugs.txt
Summary:
Now that the terminal output is truncated to max 10 issues, it's useful to have
some place where all bugs with source excerpts can be found. Since bugs.txt is
already unstructured (to the extend that it's just text), hopefully no one was
relying on it having a particular format so it's safe to change it to add
source excerpts.

Anyone wanting the old behaviour should consider parsing report.json instead,
which is intended to be more stable.

Closes https://github.com/facebook/infer/issues/520

Reviewed By: akotulski

Differential Revision: D4284175

fbshipit-source-id: 764f291