[thread-safety] don't count accesses in methods of builder classes
Summary:
Although the Builder pattern is not actually thread-safe, Builder's are not expected to be shared between threads.
Handle this by ignoring all unprotected accesses in classes the end with "Builder".
We might be able to soften this heuristic in the future by ensuring rather than assuming that Builder are not shared between methods (or, ideally, between threads).

Reviewed By: peterogithub

Differential Revision: D4280761

fbshipit-source-id: a4e6738