[infer][java] Adding the android.jar in the classpath is not needed anymore
Summary:
This is legacy code that dates back from when we needed the fields in the models to match exactly the fields in the classes being modeled. We no longer need to this. Besides, it seems that this `android.jar` stopped being part of the release for a while. So this change will not affect the results in prod

I check in details the difference on the code of Guava and it seems to remove (weird) false positives only.

Reviewed By: sblackshear

Differential Revision: D4242444

fbshipit-source-id: 84dd782