[quandary] ignore null assignments to return value in void functions
Summary:
The Java frontend translates exceptions by assigning them to the return value.
This leads to weird behavior when the return type of the function is void.
Already handled one case of this in Quandary (ignoring assignments of exceptions to return value), but was missing the case where null is assigned to the return value.
The frontend does this to "clear" the value of previously assigned exceptions.

Reviewed By: jeremydubreil

Differential Revision: D4294060

fbshipit-source-id: 6bef5ef