Use place-holders string in messages
Summary:
This diff introduces place-holders strings in error messages and evaluates them
when an error needs to be reported.

Place-holders strings are of the form %name_of_helper_function%. They will be evaluated
using and helper function that gives a value.

For example if we need to display the name of a variable in the error message we will have:

  ".... %var_name%...."

then %var_name% will be evaluated in the ast node calling the appropriate
helper function and the results will replace %var_name% in the message.

Reviewed By: dulmarod

Differential Revision: D4313133

fbshipit-source-id: bf521ca