[siof] collate multiple reports at the same line into one
Summary:
Change the domain of SIOF to be based on sets of pvar * location instead of
single pvars. This allows us to group several accesses together. However, we
still get different trace elems for different instructions in a proc. We do two
things to get around this limitation and get a trace where all accesses within
the same proc are grouped together, instead of one trace for each access:

1. A post-processing phase at the end of the analysis of one proc collects all
  the globals directly accessed in the proc into a single trace elem.
2. When creating the error trace, unpack this set into several trace elements
  to see each access (at its correct location) separately in the trace.

This is a bit hacky and another way would be to extend the API of traces to
handle in-procedure accesses natively instead of shoe-horning them. However
since SIOF is the only one to use this, it introduces less boilerplate to do it
that way for now.

Also, a few .mlis for good measure.

Reviewed By: sblackshear

Differential Revision: D4299070

fbshipit-source-id: 3bbb5c2