[log] add option to print all logs to the console
Summary:
For example: `infer --print-logs --stats -- clang -c hello.c`.

The option is not on by default.

This forwards all the output to log files to stdout or stderr as appropriate.
The multiplexing is very crude and can be improved later if needed if
stdout/err is too garbled by concurrent partial writes.

Reviewed By: jberdine

Differential Revision: D4365996

fbshipit-source-id: 7f2ab98