[java] more verbose debug output, fix for non-Buck integrations
Summary:
Currently, if we don't find `-d` or `-classes_out` on the command line then we
tell javac to redirect the compiled classes in some other directory, by default
the initial working directory. But we don't detect when these arguments are
hidden inside files (`foo` arguments on the javac command line) so the
heuristic was incomplete. Look inside these files to better tell whether we need
to make up an output directory or not.

Reviewed By: jeremydubreil

Differential Revision: D4397716

fbshipit-source-id: 30c5e4f