[cleanup] separate functions for checking annotations of return value and parameters
Summary:
`pdesc_has_annot` checks the annotations of both the return values and the parameters, which seems like a bad idea in general.
The client should have to specify which annotations they actually care about.
Converting existing uses of `pdesc_has_annot` to what I read as the intended behavior (checking the return annotation).
Will make better use of the other new functions in a follow-up.

Reviewed By: jeremydubreil

Differential Revision: D4469885

fbshipit-source-id: de5531e