[thread-safety] don't warn on accesses to fields assigned to @Functional calls
Summary:
This diff adds a set of access paths holding a value returned from a method annotated with Functional to the domain.
If a "functional" value is written to a field, we won't count that right as an unprotected access.

The idea is to be able to use the Functional annotation to get rid of benign race false positive, such as:

```
Functional T iAlwaysReturnTheSameThing();

T mCache;

T memoizedGetter() {
  if (mCache == null) {
    mCache = iAlwaysReturnTheSameThing();
  }
  return mCache;
}
```

Although there is a write-write race on `mCache`, we don't care because it will be assigned to the same value regardless of which writer wins.

Reviewed By: peterogithub

Differential Revision: D4476492

fbshipit-source-id: cfa5dfc