[infer][java] Force the option --deep when running Infer on Java Buck projects
Summary: The Buck integration assumes that the output jar file of all the dependency targets are available locally in order to retrieve the analysis from these targets. However, this is not guaranteed to be true when there is a cache hit on some targets. Running `buck build` with the option `--deep` forces this property.

Reviewed By: jvillard

Differential Revision: D4474036

fbshipit-source-id: accabfa