[quandary] improve taint propagation for unknown calls
Summary:
When the receiver type and return type of an unknown call are the same, propagate taint to both the receiver and the return type.
This does the right thing for common "builder-style" methods that both update and return the receiver.
We already had custom models for a few such methods (e.g., `StringBuilder.append`), but we can remove them now.

Reviewed By: jeremydubreil

Differential Revision: D4490071

fbshipit-source-id: 325ea88