move CLOpt.exe to Config and restrict its API, remove Interactive fake exe
Summary:
Clients should use `Config.parse_action` instead to figure in what mode they
are operating.

In particular, the biggest change is in logging. Take the `parse_action` into
account instead of the exe, and change the log/ subdirectories to be "capture",
"driver", "analyze", and "print", corresponding to the various phases of an
infer run.

Reviewed By: jberdine

Differential Revision: D4474943

fbshipit-source-id: 6d33ad3